now computing toa in ns

+++ openair1/PHY/NR_ESTIMATION/nr_ul_estimation.h
@@ -66,7 +66,7 @@ int nr_est_timing_advance_srs(const NR_DL_FRAME_PARMS *frame_parms,
 			      uint8_t N_ap,
                               const int32_t srs_estimated_channel_time[N_ap][frame_parms->ofdm_symbol_size]);
 
-float nr_est_toa_ns_srs(NR_DL_FRAME_PARMS *frame_parms,
+int32_t nr_est_toa_ns_srs(NR_DL_FRAME_PARMS *frame_parms,
 			uint8_t N_ap,
 			int32_t srs_estimated_channel_freq[N_ap][frame_parms->ofdm_symbol_size]);
